refactored to seperate calculations and ui code
