+++ ee_utils.py
@@ -1,0 +1,108 @@
+
+import pandas as pd
+import numpy as np
+import ee
+import logging
+import folium
+
+logger = logging.getLogger(__name__)
+
+
+def add_ee_layer(self, ee_image_object, vis_params, name):
+    """Adds a method for displaying Earth Engine image tiles to folium map."""
+    map_id_dict = ee.Image(ee_image_object).getMapId(vis_params)
+    folium.raster_layers.TileLayer(
+        tiles=map_id_dict["tile_fetcher"].url_format,
+        attr="Map Data &copy; <a href='https://earthengine.google.com/'>Google Earth Engine</a>",
+        name=name,
+        overlay=True,
+        control=True,
+    ).add_to(self)
+
+
+def ee_array_to_df(arr, list_of_bands):
+    """Transforms client-side ee.Image.getRegion array to pandas.DataFrame."""
+    arr = np.array(arr) # convert list to numpy array
+    df = pd.DataFrame(arr)
+
+    # Rearrange the header.
+    headers = df.iloc[0]
+    df = pd.DataFrame(df.values[1:], columns=headers)
+
+    # Convert the data to numeric values.
+    for band in list_of_bands:
+        df[band] = pd.to_numeric(df[band], errors="coerce")
+
+    # Convert the time field into a datetime.
+    df["datetime"] = pd.to_datetime(df["time"], unit="ms")
+
+    # Keep the columns of interest.
+    df = df[["time", "datetime", *list_of_bands]]
+
+    # The datetime column is defined as index.
+    df = df.set_index("datetime")
+
+    return df
+
+
+def sum_resampler(coll, freq, unit, scale_factor, band_name):
+    """
+    This function aims to resample the time scale of an ee.ImageCollection.
+    The function returns an ee.ImageCollection with the averaged sum of the
+    band on the selected frequency.
+
+    coll: (ee.ImageCollection) only one band can be handled
+    freq: (int) corresponds to the resampling frequence
+    unit: (str) corresponds to the resampling time unit.
+                must be 'day', 'month' or 'year'
+    scale_factor (float): scaling factor used to get our value in the good unit
+    band_name (str) name of the output band
+    """
+
+    # Define initial and final dates of the collection.
+    firstdate = ee.Date(
+        coll.sort("system:time_start", True).first().get("system:time_start")
+    )
+
+    lastdate = ee.Date(
+        coll.sort("system:time_start", False).first().get("system:time_start")
+    )
+
+    # Calculate the time difference between both dates.
+    # https://developers.google.com/earth-engine/apidocs/ee-date-difference
+    diff_dates = lastdate.difference(firstdate, unit)
+
+    # Define a new time index (for output).
+    new_index = ee.List.sequence(0, ee.Number(diff_dates), freq)
+
+    # Define the function that will be applied to our new time index.
+    def apply_resampling(date_index):
+        # Define the starting date to take into account.
+        startdate = firstdate.advance(ee.Number(date_index), unit)
+
+        # Define the ending date to take into account according
+        # to the desired frequency.
+        enddate = firstdate.advance(ee.Number(date_index).add(freq), unit)
+
+        # Calculate the number of days between starting and ending days.
+        diff_days = enddate.difference(startdate, "day")
+
+        # Calculate the composite image.
+        image = (
+            coll.filterDate(startdate, enddate)
+            .mean()
+            .multiply(diff_days)
+            .multiply(scale_factor)
+            .rename(band_name)
+        )
+
+        # Return the final image with the appropriate time index.
+        return image.set("system:time_start", startdate.millis())
+
+    # Map the function to the new time index.
+    res = new_index.map(apply_resampling)
+
+    # Transform the result into an ee.ImageCollection.
+    res = ee.ImageCollection(res)
+
+    return res
